Split processing out into its own class. WIP
